add verify-hostname command-line option
This option wasn't working since the v0.8 refactor. We're now
implementing it, however changing its default value to false, at least
at this moment.

HTTP only. TCP services are not properly using the list of hosts from
the TLS field, we'll leave TCP verification when this is properly fixed.

This is a breaking change if one has `--verify-hostname` configured in
the controller bootstrap - it wasn't working before and now it'll start
to block certificates that doesn't match the hostnames.